UTC <-> TAI Conversion
A number of tests had to be adjusted after leap second calculations were added, because they hadn't previously accounted for leap seconds being in use. A couple of other adjustments to the actual code were needed to allow things to actually function as expected, especially in the Unix calendar implementation. There's still plenty to do, but this is a feature the PHP implementation never had, so I'm pretty happy with it. :smile: